sstables: leveled_manifest: prevent benign precision loss warning
Casting from the maximum int64_t to double loses precision, because
int64_t has 64 bits of precision while double has only 53. Clang
warns about it. Since it's not a real problem here, add an explicit
cast to silence the warning.

+++ sstables/leveled_manifest.hh
@@ -151,7 +151,7 @@ public:
             return 4L * max_sstable_size_in_bytes;
         }
         double bytes = pow(leveled_fan_out, level) * max_sstable_size_in_bytes;
-        if (bytes > std::numeric_limits<int64_t>::max()) {
+        if (bytes > double(std::numeric_limits<int64_t>::max())) {
             throw std::runtime_error(format("At most {:d} bytes may be in a compaction level; your maxSSTableSize must be absurdly high to compute {:f}", 
                 std::numeric_limits<int64_t>::max(), bytes));
         }
